Refinements

+++ startup.sh
@@ -4,7 +4,14 @@
 cd jbrowse2
 rm -rf test_data
 
+# check to make sure data directory exists; easier to bail now if not
+if [ ! -e "$SERVICE_FILES_MOUNT" ]; then
+    echo "$SERVICE_FILES_MOUNT does not exist. Is the volume configured?"
+    exit 1
+fi
+
 # link to mounted track data
+rm -f data
 ln -s $SERVICE_FILES_MOUNT data
 
 # need to copy the config.json file and fix paths
